Add files via upload
The objective of this project is to develop a comprehensive system for object recognition in videos using the YOLO (You Only Look Once) algorithm and text extraction from videos using Optical Character Recognition (OCR) techniques. The system aims to extract both detected objects and extracted text as features and utilize them to train a multi-class classification model for video classification.